% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{A take on a children's game.}
\usage{
FizzBuzz(x)
}
\arguments{
\item{x}{a vector}
}
\value{
First \code{x} of the game Fizz Buzz
}
\description{
Fizz Buzz is a children's game. The FizzBuzz function will output a vector
of the first x terms given as a parameter.
}
\details{
For any number evenly divisible by 3, “Fizz” will appear instead, and any number
evenly divisible by 5, “Buzz” will appear, and if it is divisible by
both, you'll see “Fizz Buzz.”
}
\examples{
FizzBuzz(1:30)
FizzBuzz(15:45)
}
