% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of weather information at Flagstaff’s Pulliam Airport from 1950 to 2019.}
\format{
a data frame with 22967 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{RAIN}{Daily precipitation/rainfall in inches.}
  \item{SNOW}{Daily snowfall in inches.}
  \item{TMAX}{Daily maximum temperature in degrees Farhenheit.}
  \item{TMIN}{Daily minimum temperature in degrees Farhenheit.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of weather information at Flagstaff’s Pulliam Airport from 1950 to 2019.
}
\keyword{datasets}
